+++ src/test/java/roomescape/repository/ReservationTimeRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.DynamicTest.dynamicTest;
@@ -24,7 +24,7 @@ import roomescape.domain.ReservationTime;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class ReservationTimeDaoTest {
+public class ReservationTimeRepositoryTest {
 
     @LocalServerPort
     int serverPort;
